refactor all remote staging into their own files
